# bpf/proxy: do not update unchanged services