## Description

before:

```
BenchmarkServiceUpdate/Services_1_Endpoints_1_mockMaps_true-12         	  162416	      7359 ns/op	    2854 B/op	      29 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_1_mockMaps_false-12        	   81061	     15542 ns/op	    3465 B/op	      58 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_1_mockMaps_true_+_derived-12         	  138472	      9048 ns/op	    3258 B/op	      42 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_1_mockMaps_false_+_derived-12        	   65394	     21141 ns/op	    4201 B/op	      86 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_10_mockMaps_true-12                  	   97296	     11597 ns/op	    3646 B/op	      88 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_10_mockMaps_false-12                 	   23743	     59862 ns/op	    7812 B/op	     286 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_10_mockMaps_true_+_derived-12        	   94364	     13713 ns/op	    4058 B/op	     101 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_10_mockMaps_false_+_derived-12       	   22196	     67150 ns/op	    8744 B/op	     322 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_1_mockMaps_true-12                  	   54650	     24844 ns/op	    6592 B/op	     137 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_1_mockMaps_false-12                 	   17290	    106666 ns/op	   14427 B/op	     439 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_1_mockMaps_true_+_derived-12        	   33363	     44043 ns/op	   16546 B/op	     246 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_1_mockMaps_false_+_derived-12       	   10000	    104883 ns/op	   18782 B/op	     454 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_10_mockMaps_true-12                 	   20980	     80944 ns/op	   17695 B/op	     840 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_10_mockMaps_false-12                	    4599	    315012 ns/op	   34078 B/op	    1540 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_10_mockMaps_true_+_derived-12       	   15526	    119980 ns/op	   31465 B/op	    1053 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_10_mockMaps_false_+_derived-12      	    3004	    372549 ns/op	   43812 B/op	    1721 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_1_mockMaps_true-12                 	    6814	    334825 ns/op	   81528 B/op	    1796 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_1_mockMaps_false-12                	    2146	    660768 ns/op	   81352 B/op	    2632 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_1_mockMaps_true_+_derived-12       	    3351	    360457 ns/op	  112309 B/op	    1729 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_1_mockMaps_false_+_derived-12      	    1080	    969643 ns/op	  166891 B/op	    4142 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_10_mockMaps_true-12                	    2587	    542724 ns/op	  110076 B/op	    5532 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_10_mockMaps_false-12               	     430	   3359818 ns/op	  312542 B/op	   14371 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_10_mockMaps_true_+_derived-12      	    1594	    749268 ns/op	  177226 B/op	    6237 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_10_mockMaps_false_+_derived-12     	     324	   3614548 ns/op	  397823 B/op	   15874 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_1_mockMaps_true-12                	     604	   2086942 ns/op	  579936 B/op	   10020 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_1_mockMaps_false-12               	     157	   6931701 ns/op	  939961 B/op	   26020 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_1_mockMaps_true_+_derived-12      	     326	   3641477 ns/op	 1402845 B/op	   17057 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_1_mockMaps_false_+_derived-12     	     116	   9721405 ns/op	 1946979 B/op	   41056 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_10_mockMaps_true-12               	     214	   5358566 ns/op	 1227952 B/op	   55023 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_10_mockMaps_false-12              	      39	  33622087 ns/op	 3244300 B/op	  143037 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_10_mockMaps_true_+_derived-12     	     176	   6786472 ns/op	 2051127 B/op	   62058 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_10_mockMaps_false_+_derived-12    	      30	  35768808 ns/op	 4253767 B/op	  158129 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_1_mockMaps_true-12               	      58	  22102178 ns/op	 4824745 B/op	  100022 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_1_mockMaps_false-12              	      16	  73805093 ns/op	 8424858 B/op	  260019 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_1_mockMaps_true_+_derived-12     	      31	  38136083 ns/op	11935406 B/op	  170311 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_1_mockMaps_false_+_derived-12    	       9	 111704593 ns/op	17386875 B/op	  410344 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_10_mockMaps_true-12              	      20	  60660517 ns/op	11306271 B/op	  550023 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_10_mockMaps_false-12             	       3	 355060990 ns/op	31465437 B/op	 1430075 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_10_mockMaps_true_+_derived-12    	      15	  79439345 ns/op	18422458 B/op	  620336 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_10_mockMaps_false_+_derived-12   	       3	 367723247 ns/op	40420541 B/op	 1580352 allocs/op
```

after:

```
BenchmarkServiceUpdate/Services_1_Endpoints_1_mockMaps_true-12         	  176314	      7559 ns/op	    2848 B/op	      29 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_1_mockMaps_false-12        	   78752	     15063 ns/op	    3408 B/op	      54 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_1_mockMaps_true_+_derived-12         	  138424	      8882 ns/op	    3247 B/op	      41 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_1_mockMaps_false_+_derived-12        	   64789	     19201 ns/op	    4094 B/op	      80 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_10_mockMaps_true-12                  	   98176	     11411 ns/op	    3568 B/op	      83 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_10_mockMaps_false-12                 	   25051	     46003 ns/op	    6640 B/op	     225 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_10_mockMaps_true_+_derived-12        	   96440	     13263 ns/op	    3973 B/op	      95 allocs/op
BenchmarkServiceUpdate/Services_1_Endpoints_10_mockMaps_false_+_derived-12       	   23035	     52635 ns/op	    7392 B/op	     252 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_1_mockMaps_true-12                  	   69788	     17477 ns/op	    5226 B/op	      29 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_1_mockMaps_false-12                 	   48255	     25608 ns/op	    6032 B/op	      55 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_1_mockMaps_true_+_derived-12        	   43879	     35228 ns/op	   13950 B/op	      76 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_1_mockMaps_false_+_derived-12       	   28060	     50083 ns/op	   16154 B/op	     119 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_10_mockMaps_true-12                 	   47505	     28737 ns/op	    6206 B/op	      83 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_10_mockMaps_false-12                	   18351	     73060 ns/op	   10560 B/op	     226 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_10_mockMaps_true_+_derived-12       	   27679	     49572 ns/op	   16093 B/op	     134 allocs/op
BenchmarkServiceUpdate/Services_10_Endpoints_10_mockMaps_false_+_derived-12      	   12722	    101302 ns/op	   23997 B/op	     303 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_1_mockMaps_true-12                 	   10000	    105758 ns/op	   34130 B/op	      33 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_1_mockMaps_false-12                	   10000	    120311 ns/op	   34691 B/op	      59 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_1_mockMaps_true_+_derived-12       	    5769	    353167 ns/op	  157594 B/op	     542 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_1_mockMaps_false_+_derived-12      	    5979	    227843 ns/op	   98097 B/op	     383 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_10_mockMaps_true-12                	    7845	    161347 ns/op	   34840 B/op	      86 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_10_mockMaps_false-12               	    6801	    205133 ns/op	   37863 B/op	     227 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_10_mockMaps_true_+_derived-12      	    4336	    276856 ns/op	   97997 B/op	     398 allocs/op
BenchmarkServiceUpdate/Services_100_Endpoints_10_mockMaps_false_+_derived-12     	    3667	    335963 ns/op	  101334 B/op	     554 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_1_mockMaps_true-12                	    1110	   1163689 ns/op	  468045 B/op	      29 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_1_mockMaps_false-12               	    1137	   1037475 ns/op	  468555 B/op	      53 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_1_mockMaps_true_+_derived-12      	     552	   2584943 ns/op	 1250671 B/op	    3073 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_1_mockMaps_false_+_derived-12     	     519	   2252204 ns/op	 1251467 B/op	    3112 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_10_mockMaps_true-12               	     658	   1725658 ns/op	  468642 B/op	      74 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_10_mockMaps_false-12              	     650	   1613262 ns/op	  471240 B/op	     195 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_10_mockMaps_true_+_derived-12     	     370	   2781708 ns/op	 1251768 B/op	    3127 allocs/op
BenchmarkServiceUpdate/Services_1000_Endpoints_10_mockMaps_false_+_derived-12    	     441	   2790953 ns/op	 1254930 B/op	    3283 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_1_mockMaps_true-12               	      84	  12613745 ns/op	 3705096 B/op	      32 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_1_mockMaps_false-12              	      98	  12628490 ns/op	 3705673 B/op	      58 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_1_mockMaps_true_+_derived-12     	      46	  27247028 ns/op	10417729 B/op	   30336 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_1_mockMaps_false_+_derived-12    	      49	  29184740 ns/op	10417524 B/op	   30374 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_10_mockMaps_true-12              	      62	  21360122 ns/op	 3705309 B/op	      85 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_10_mockMaps_false-12             	      62	  22940588 ns/op	 3708641 B/op	     226 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_10_mockMaps_true_+_derived-12    	      36	  35517080 ns/op	10422811 B/op	   30395 allocs/op
BenchmarkServiceUpdate/Services_10000_Endpoints_10_mockMaps_false_+_derived-12   	      37	  37341294 ns/op	10421588 B/op	   30546 allocs/op
```
<!-- A few sentences describing the overall goals of the pull request's commits.
Please include
- the type of fix - (e.g. bug fix, new feature, documentation)
- some details on _why_ this PR should be merged
- the details of the testing you've done on it (both manual and automated)
- which components are affected by this PR
-->

## Todos
- [ ] Unit tests (full coverage)
- [ ] Integration tests (delete as appropriate) In plan/Not needed/Done
- [ ] Documentation
- [ ] Backport
- [ ] Release note

## Release Note
<!-- Writing a release note:
- By default, no release note action is required.
- If you're unsure whether or not your PR needs a note, ask your reviewer for guidance.
- If this PR requires a release note, update the block below to include a concise note describing
  the change and any important impacts this PR may have.
-->

```release-note
None required
```